half work was done.html
In general, the login and registration window was added, after that I added icons to each field, checkbox, forgot password, don’t have an account, already have an account, i agree to the terms, register links was added. Buttons "Home, about, contact" added. It was all related to CSS and js. most of the work was done left to tie up the steel work